+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Fixed
+- Fix bug where default properties "name" and "message" on error objects were
+  not reported by `Honeybadger.notify`.
 
 ## [0.4.4] - 2017-02-23
 ### Added